requirements.txt
Django==5.0.2
python-dotenv==1.0.1
requests==2.31.0
langchain==0.1.12
langchain-core==0.1.32
langchain-openai==0.0.8
langgraph==0.0.26
openai==1.14.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
numpy==1.26.4
scikit-learn==1.4.1.post1
